Porting patch for Mac OS X 10.3, implements the following changes:
  - Initialize some variables the 3.3 optimizer loses track of.
  - Add some type conversions to ensure values of time_t get passed correctly.
  - Make message for GLOB_NOMATCH optional because Panther doesn't know it.
  - Change ruby binary path to tuneable, testing on Mac OS X 10.3 does
    require calling make like this:
      make RUBY=ruby1.8 test

+++ src/libgtcore/timer.c
@@ -65,9 +65,9 @@ void timer_show(Timer *t, FILE *fp)
     timer_stop(t);
   assert(t->state == TIMER_STOPPED);
   fprintf(fp, "%lds real %lds user %lds system\n",
-          t->stop_tv.tv_sec - t->start_tv.tv_sec,
-          t->stop_ru.ru_utime.tv_sec - t->stop_ru.ru_utime.tv_sec,
-          t->stop_ru.ru_stime.tv_sec - t->stop_ru.ru_stime.tv_sec);
+          (long)(t->stop_tv.tv_sec - t->start_tv.tv_sec),
+          (long)(t->stop_ru.ru_utime.tv_sec - t->stop_ru.ru_utime.tv_sec),
+          (long)(t->stop_ru.ru_stime.tv_sec - t->stop_ru.ru_stime.tv_sec));
 }
 
 void timer_del(Timer *t)
